[refactor] SpaceMember dirctory sturcture refactoring
