+++ src/test/java/space/space_spring/domain/pay/application/service/CreatePayServiceTest.java
@@ -9,7 +9,7 @@ import space.space_spring.domain.pay.application.port.in.CreatePayCommand;
 import space.space_spring.domain.pay.application.port.out.CreatePayPort;
 import space.space_spring.domain.pay.application.port.out.LoadSpaceMemberPort;
 import space.space_spring.domain.space.domain.Space;
-import space.space_spring.domain.spaceMember.SpaceMember;
+import space.space_spring.domain.spaceMember.domian.SpaceMember;
 import space.space_spring.domain.user.User;
 import space.space_spring.global.exception.CustomException;
